refactor: project structure and add core game components
