+++ src/ai/__init__.py
@@ -1,0 +1,6 @@
+# AI Package
+"""Neural network implementation for AI-controlled asteroids."""
+
+from src.ai.neural_network import neural_network
+
+__all__ = ['neural_network']
