+++ src/core/__init__.py
@@ -1,0 +1,8 @@
+# Core Game Package
+"""Core game logic, constants, and game states."""
+
+from src.core.game_states import GAME_STATES
+from src.core import constants
+from src.core import game
+
+__all__ = ['GAME_STATES', 'constants', 'game']
