+++ src/entities/__init__.py
@@ -1,0 +1,8 @@
+# Game Entities Package
+"""Game entities: rocket, asteroid, and bullet classes."""
+
+from src.entities.rocket import rocket, ROCKET_STATUS
+from src.entities.asteroid import asteroid, ASTEROID_STATUS
+from src.entities.bullet import bullet
+
+__all__ = ['rocket', 'ROCKET_STATUS', 'asteroid', 'ASTEROID_STATUS', 'bullet']
