+++ src/entities/rocket.py
@@ -5,12 +5,13 @@ Defines the rocket class
 import numpy as np
 import pygame
 import enum
-from constants import window_width
-from constants import window_height
-from constants import generalise_height
-from constants import rocket_accel_coeff
-from constants import rocket_velocity_coeff
-import bullet
+from src.core.constants import window_width
+from src.core.constants import window_height
+from src.core.constants import generalise_height
+from src.core.constants import rocket_accel_coeff
+from src.core.constants import rocket_velocity_coeff
+from src.core.constants import get_asset_path
+from src.entities import bullet
 
 
 class ROCKET_STATUS(enum.Enum):
@@ -37,8 +38,9 @@ class rocket(pygame.sprite.Sprite):
         self.velocity = np.array([0.0, 0.0])
         self.angle = 0.0
         self.status = ROCKET_STATUS.ALIVE
+        asset_path = get_asset_path('images', 'rocket2.png')
         self.unrotated_image = pygame.transform.smoothscale(pygame.image.load(
-            'rocket2.png').convert_alpha(),
+            asset_path).convert_alpha(),
             (generalise_height(20), generalise_height(20)))
         self.last_shot_tick = -1000
