+++ src/ui/__init__.py
@@ -1,0 +1,6 @@
+# UI Package
+"""User interface components: menus and UI elements."""
+
+from src.ui import main_menu
+
+__all__ = ['main_menu']
